requirements.txt
discord.py>=1.0.0a
toml>=0.10.0
